+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
 ### 7.4.14 (07/13/2020 - )
+* redis: change expiration precision from second to milliseconds
 
 ### 7.4.13 (07/09/2020 - 07/13/2020)
 * cache: optimize local cache performance